Start rewriting code to get rid of old parser.
