+++ Cargo.toml
@@ -1,5 +1,5 @@
-# Copyright Jeron Lau 2018-2019.
-# Copyright Dylan Ede 2016.
+# Copyright Jeron Aldaron Lau 2018-2019.
+#
 # Dual-licensed under either the MIT License or the Boost Software License, Version 1.0.
 # (See accompanying file LICENSE_1_0.txt or copy at https://www.boost.org/LICENSE_1_0.txt)
 
@@ -27,13 +27,16 @@ exclude = [
 ]
 
 [dependencies]
-footile = "0.3"
-byteorder = "1.3"
-unicode-normalization = "0.1"
-unicode-width = "0.1"
+footile = { path = "../../fork/footile" } # "0.3"
+# byteorder = "1.3"
+# unicode-normalization = "0.1"
+# unicode-width = "0.1"
+ttf-parser = "0.2"
 
 [dev-dependencies]
 svg = "0.5"
+footile = { path = "../../fork/footile" } # "0.3"
+png_pong = "0.0.2"
 
 [features]
 default = ["builtin-font"]
