+++ LICENSE
@@ -1,7 +1,6 @@
 MIT License
 
-Copyright (c) 2018-2019 Jeron Lau
-Copyright (c) 2016 Dylan Ede
+Copyright (c) 2018-2019 Jeron Aldaron Lau
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
