+++ LICENSE_1_0.txt
@@ -1,5 +1,7 @@
 Boost Software License - Version 1.0 - August 17th, 2003
 
+Copyright (c) 2018-2019 Jeron Aldaron Lau
+
 Permission is hereby granted, free of charge, to any person or organization
 obtaining a copy of the software and accompanying documentation covered by
 this license (the "Software") to use, reproduce, display, distribute,
