+++ examples/raster.rs
@@ -12,10 +12,6 @@ fn main() {
     let mut p = Plotter::new(2048, 2048);
     let mut r = Raster::new(p.width(), p.height());
 
-    //    let path = font.render(STR, (1024.0, 0.0), (FONT_SIZE, FONT_SIZE)).vertical().right_to_left();
-    //    r.over(p.fill(path, FillRule::NonZero), Rgba8::rgb(0, 0, 0));
-    //    let path = font.render(STR, (1024.0, 0.0), (FONT_SIZE, FONT_SIZE)).vertical();
-    //    r.over(p.fill(path, FillRule::NonZero), Rgba8::rgb(0, 0, 0));
     let mut path = font.render(STR, (0.0, 0.0), (FONT_SIZE, FONT_SIZE));
     r.over(p.fill(&mut path, FillRule::NonZero), Rgba8::rgb(0, 0, 0));
     let mut path = font
